Switch to using injectable to setup route for metrics
