+++ src/main/routes/metrics/metrics-query.ts
@@ -2,6 +2,6 @@
  * Copyright (c) OpenLens Authors. All rights reserved.
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
-
-export * from "./kubeconfig-route";
-export * from "./version-route";
+export type IMetricsQuery = string | string[] | {
+  [metricName: string]: string;
+};
